components/sensor-simulator/requirements.txt
awsiotsdk>=1.19.0